+++ pulp_container/app/authorization.py
@@ -7,22 +7,17 @@ import uuid
 import jwt
 
 from datetime import datetime
-from aiohttp import web
 
 from django.conf import settings
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
 
-from pulpcore.plugin.content import Handler
-
 TOKEN_EXPIRATION_TIME = 300
 
-KNOWN_SERVICES = [re.sub(r'(http://|https://)', '', settings.CONTENT_ORIGIN, count=1)]
-ANONYMOUS_USER = ''
-EMPTY_ACCESS_SCOPE = '::'
+KNOWN_SERVICES = [re.sub(r"(http://|https://)", "", settings.CONTENT_ORIGIN, count=1)]
 
 
-class AuthorizationService(Handler):
+class AuthorizationService:
     """
     A class responsible for generating and managing a Bearer token.
 
@@ -30,7 +25,7 @@ class AuthorizationService(Handler):
     according to a user's scope.
     """
 
-    async def generate_token(self, request):
+    def generate_token(self, username, service, scope):
         """
         Generate a Bearer token.
 
@@ -38,54 +33,55 @@ class AuthorizationService(Handler):
         adjusted according the documentation https://docs.docker.com/registry/spec/auth/jwt/.
 
         Args:
-            request(:class:`~aiohttp.web.Request`): The request to prepare a response for.
+            username (str): Requesting user.
+            service (str): Name of the service access is granted to.
+            scope (str): Scope of the resource that is to be accessed.
 
         Returns:
-            class:`aiohttp.web_response.Response`: A newly generated Bearer token.
+            dict: A newly generated Bearer token.
 
         """
-        with open(settings.PUBLIC_KEY_PATH, 'rb') as public_key:
+        with open(settings.PUBLIC_KEY_PATH, "rb") as public_key:
             kid = self.generate_kid_header(public_key.read())
 
         current_datetime = datetime.now()
 
-        token_queries = TokenRequestQueries.init_from(request)
-        access = self.determine_access(ANONYMOUS_USER, token_queries.scope)
-        token_server = getattr(settings, 'TOKEN_SERVER', '')
+        access = self.determine_access(username, scope)
+        token_server = getattr(settings, "TOKEN_SERVER", "")
         claim_set = self._generate_claim_set(
             access=[access],
-            audience=token_queries.service,
+            audience=service,
             issued_at=int(current_datetime.timestamp()),
             issuer=token_server,
-            subject=ANONYMOUS_USER
+            subject=username,
         )
 
-        with open(settings.PRIVATE_KEY_PATH, 'rb') as private_key:
+        with open(settings.PRIVATE_KEY_PATH, "rb") as private_key:
             binary_token = jwt.encode(
-                claim_set, private_key.read(),
+                claim_set,
+                private_key.read(),
                 algorithm=settings.TOKEN_SIGNATURE_ALGORITHM,
-                headers={'kid': kid}
+                headers={"kid": kid},
             )
-        token = binary_token.decode('utf8')
-        current_datetime_utc = current_datetime.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
-        return web.json_response({
-            'expires_in': TOKEN_EXPIRATION_TIME,
-            'issued_at': current_datetime_utc,
-            'token': token
-        })
+        token = binary_token.decode("utf8")
+        current_datetime_utc = current_datetime.strftime("%Y-%m-%dT%H:%M:%S.%fZ")
+        return {
+            "expires_in": TOKEN_EXPIRATION_TIME,
+            "issued_at": current_datetime_utc,
+            "token": token,
+        }
 
     def generate_kid_header(self, public_key):
         """Generate kid header in a libtrust compatible format."""
         decoded_key = self._convert_key_format_from_pem_to_der(public_key)
-        truncated_sha256 = hashlib.sha256(decoded_key).hexdigest()[:30].encode('utf8')
-        encoded_base32 = base64.b32encode(truncated_sha256).decode('utf8')
+        truncated_sha256 = hashlib.sha256(decoded_key).hexdigest()[:30].encode("utf8")
+        encoded_base32 = base64.b32encode(truncated_sha256).decode("utf8")
         return self._split_into_encoded_groups(encoded_base32)
 
     def _convert_key_format_from_pem_to_der(self, public_key):
         key_in_pem_format = serialization.load_pem_public_key(public_key, default_backend())
         key_in_der_format = key_in_pem_format.public_bytes(
-            serialization.Encoding.DER,
-            serialization.PublicFormat.SubjectPublicKeyInfo
+            serialization.Encoding.DER, serialization.PublicFormat.SubjectPublicKeyInfo
         )
         return key_in_der_format
 
@@ -94,7 +90,7 @@ class AuthorizationService(Handler):
         kid = encoded_base32[:4]
         for index, char in enumerate(encoded_base32[4:], start=0):
             if index % 4 == 0:
-                kid += ':' + char
+                kid += ":" + char
             else:
                 kid += char
         return kid
@@ -116,64 +112,24 @@ class AuthorizationService(Handler):
             list: An intersected set of the requested and the allowed access.
 
         """
-        typ, name, actions = scope.split(':')
-        actions_list = actions.split(',')
-        permitted_actions = list(set(actions_list).intersection(['pull']))
-        return {'type': typ, 'name': name, 'actions': permitted_actions}
+        typ, name, actions = scope.split(":")
+        actions_list = actions.split(",")
+        permissions = {"pull"}
+        if user == "admin":
+            permissions.add("push")
+        permitted_actions = list(set(actions_list).intersection(permissions))
+        return {"type": typ, "name": name, "actions": permitted_actions}
 
     def _generate_claim_set(self, issuer, issued_at, subject, audience, access):
         token_id = str(uuid.UUID(int=random.getrandbits(128), version=4))
         expiration = issued_at + TOKEN_EXPIRATION_TIME
         return {
-            'access': access,
-            'aud': audience,
-            'exp': expiration,
-            'iat': issued_at,
-            'iss': issuer,
-            'jti': token_id,
-            'nbf': issued_at,
-            'sub': subject
+            "access": access,
+            "aud": audience,
+            "exp": expiration,
+            "iat": issued_at,
+            "iss": issuer,
+            "jti": token_id,
+            "nbf": issued_at,
+            "sub": subject,
         }
-
-
-class TokenRequestQueries:
-    """A data class that holds data retrieved from the request queries."""
-
-    def __init__(self, scope, service):
-        """
-        Store a scope and a service.
-
-        Args:
-            scope (str): A requested scope.
-            service (str): A service that request a Bearer token.
-
-        """
-        self.scope = scope
-        self.service = service
-
-    @classmethod
-    def init_from(cls, request):
-        """
-        Initialize the actual class with data retrieved from the request queries.
-
-        In this method, a validity and a presence of required queries (scope, service)
-        is checked as well. If the scope is not specified, the method checks if a user
-        is trying to access root endpoint only. Then, the scope is not relevant anymore
-        and initialized to empty type, name, and requested actions ('::').
-        """
-        try:
-            scope = request.query['scope']
-        except KeyError:
-            if request.match_info:
-                raise web.HTTPBadRequest(reason='A scope was not provided.')
-            else:
-                scope = EMPTY_ACCESS_SCOPE
-
-        try:
-            service = request.query['service']
-        except KeyError:
-            raise web.HTTPBadRequest(reason='A service name was not provided.')
-        if service not in KNOWN_SERVICES:
-            raise web.HTTPBadRequest(reason='A provided service is unknown.')
-
-        return cls(scope, service)