+++ pulp_container/app/tasks/untag.py
@@ -8,14 +8,9 @@ def untag_image(tag, repository_pk):
     repository = ContainerRepository.objects.get(pk=repository_pk)
     latest_version = repository.latest_version()
 
-    tags_in_latest_repository = latest_version.content.filter(
-        pulp_type="container.tag"
-    )
+    tags_in_latest_repository = latest_version.content.filter(pulp_type="container.tag")
 
-    tags_to_remove = Tag.objects.filter(
-        pk__in=tags_in_latest_repository,
-        name=tag
-    )
+    tags_to_remove = Tag.objects.filter(pk__in=tags_in_latest_repository, name=tag)
 
     with repository.new_version() as repository_version:
         repository_version.remove_content(tags_to_remove)