+++ pulp_container/tests/functional/api/test_pull_content.py
@@ -1,10 +1,11 @@
 # coding=utf-8
 """Tests that verify that images served by Pulp can be pulled."""
 import contextlib
+import requests
 import unittest
 from urllib.parse import urljoin
 
-from pulp_smash import cli, config, exceptions
+from pulp_smash import api, cli, config, exceptions
 from pulp_smash.pulp3.utils import (
     delete_orphans,
     get_content,
@@ -16,15 +17,17 @@ from pulp_container.tests.functional.utils import (
     core_client,
     gen_container_client,
     gen_container_remote,
-    gen_token_signing_keys,
     get_docker_hub_remote_blobsums,
     monitor_task,
+    BearerTokenAuth,
+    AuthenticationHeaderQueries,
 )
 from pulp_container.tests.functional.constants import (
     CONTAINER_CONTENT_NAME,
     REPO_UPSTREAM_NAME,
     REPO_UPSTREAM_TAG,
 )
+from pulp_container.constants import MEDIA_TYPE
 
 from pulpcore.client.pulp_container import (
     ContainerContainerDistribution,
@@ -55,8 +58,8 @@ class PullContentTestCase(unittest.TestCase):
         * `Pulp #4460 <https://pulp.plan.io/issues/4460>`_
         """
         cls.cfg = config.get_config()
-        gen_token_signing_keys(cls.cfg)
 
+        cls.client = api.Client(cls.cfg, api.code_handler)
         client_api = gen_container_client()
         cls.repositories_api = RepositoriesContainerApi(client_api)
         cls.remotes_api = RemotesContainerApi(client_api)
@@ -84,9 +87,7 @@ class PullContentTestCase(unittest.TestCase):
 
             # Step 4.
             distribution_response = cls.distributions_api.create(
-                ContainerContainerDistribution(
-                    **gen_distribution(repository=cls.repo.pulp_href)
-                )
+                ContainerContainerDistribution(**gen_distribution(repository=cls.repo.pulp_href))
             )
             created_resources = monitor_task(distribution_response.task)
             distribution = cls.distributions_api.read(created_resources[0])
@@ -105,7 +106,7 @@ class PullContentTestCase(unittest.TestCase):
             distribution = cls.distributions_api.read(created_resources[0])
             cls.distribution_with_repo_version = cls.distributions_api.read(distribution.pulp_href)
             cls.teardown_cleanups.append(
-                (cls.distributions_api.delete, cls.distribution_with_repo_version.pulp_href)
+                (cls.distributions_api.delete, cls.distribution_with_repo_version.pulp_href,)
             )
 
             # remove callback if everything goes well
@@ -126,21 +127,42 @@ class PullContentTestCase(unittest.TestCase):
         """
         # Get local checksums for content synced from remote registy
         checksums = [
-            content["digest"] for content
-            in get_content(self.repo.to_dict())[CONTAINER_CONTENT_NAME]
+            content["digest"]
+            for content in get_content(self.repo.to_dict())[CONTAINER_CONTENT_NAME]
         ]
 
         # Assert that at least one layer is synced from remote:latest
         # and the checksum matched with remote
         self.assertTrue(
-            any(
-                [
-                    result["blobSum"] in checksums
-                    for result in get_docker_hub_remote_blobsums()
-                ]
-            ),
-            "Cannot find a matching layer on remote registry."
+            any([result["blobSum"] in checksums for result in get_docker_hub_remote_blobsums()]),
+            "Cannot find a matching layer on remote registry.",
         )
+
+    def test_api_performes_schema_conversion(self):
+        """Verify pull via token with accepted content type.
+        """
+        image_path = "/v2/{}/manifests/{}".format(self.distribution_with_repo.base_path, "latest")
+        latest_image_url = urljoin(self.cfg.get_base_url(), image_path)
+
+        with self.assertRaises(requests.HTTPError) as cm:
+            self.client.get(latest_image_url, headers={"Accept": MEDIA_TYPE.MANIFEST_V1})
+
+        content_response = cm.exception.response
+        self.assertEqual(content_response.status_code, 401)
+
+        authenticate_header = content_response.headers["Www-Authenticate"]
+        queries = AuthenticationHeaderQueries(authenticate_header)
+        content_response = self.client.get(
+            queries.realm, params={"service": queries.service, "scope": queries.scope}
+        )
+        token = content_response.json()["token"]
+        content_response = self.client.get(
+            latest_image_url,
+            auth=BearerTokenAuth(token),
+            headers={"Accept": MEDIA_TYPE.MANIFEST_V1},
+        )
+        base_content_type = content_response.headers["Content-Type"].split(";")[0]
+        self.assertIn(base_content_type, {MEDIA_TYPE.MANIFEST_V1, MEDIA_TYPE.MANIFEST_V1_SIGNED})
 
     def test_pull_image_from_repository(self):
         """Verify that a client can pull the image from Pulp.
@@ -151,14 +173,9 @@ class PullContentTestCase(unittest.TestCase):
         4. Ensure image is deleted after the test.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            self.distribution_with_repo.base_path
-        )
+        local_url = urljoin(self.cfg.get_base_url(), self.distribution_with_repo.base_path)
 
         registry.pull(local_url)
         self.teardown_cleanups.append((registry.rmi, local_url))
@@ -167,10 +184,7 @@ class PullContentTestCase(unittest.TestCase):
         registry.pull(REPO_UPSTREAM_NAME)
         remote_image = registry.inspect(REPO_UPSTREAM_NAME)
 
-        self.assertEqual(
-            local_image[0]["Id"],
-            remote_image[0]["Id"]
-        )
+        self.assertEqual(local_image[0]["Id"], remote_image[0]["Id"])
         registry.rmi(REPO_UPSTREAM_NAME)
 
     def test_pull_image_from_repository_version(self):
@@ -182,14 +196,9 @@ class PullContentTestCase(unittest.TestCase):
         4. Ensure image is deleted after the test.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            self.distribution_with_repo_version.base_path
-        )
+        local_url = urljoin(self.cfg.get_base_url(), self.distribution_with_repo_version.base_path)
 
         registry.pull(local_url)
         self.teardown_cleanups.append((registry.rmi, local_url))
@@ -198,10 +207,7 @@ class PullContentTestCase(unittest.TestCase):
         registry.pull(REPO_UPSTREAM_NAME)
         remote_image = registry.inspect(REPO_UPSTREAM_NAME)
 
-        self.assertEqual(
-            local_image[0]["Id"],
-            remote_image[0]["Id"]
-        )
+        self.assertEqual(local_image[0]["Id"], remote_image[0]["Id"])
         registry.rmi(REPO_UPSTREAM_NAME)
 
     def test_pull_image_with_tag(self):
@@ -213,31 +219,22 @@ class PullContentTestCase(unittest.TestCase):
         4. Ensure image is deleted after the test.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            self.distribution_with_repo.base_path
-        ) + REPO_UPSTREAM_TAG
+        local_url = (
+            urljoin(self.cfg.get_base_url(), self.distribution_with_repo.base_path)
+            + REPO_UPSTREAM_TAG
+        )
 
         registry.pull(local_url)
         self.teardown_cleanups.append((registry.rmi, local_url))
         local_image = registry.inspect(local_url)
 
         registry.pull(REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG)
-        self.teardown_cleanups.append(
-            (registry.rmi, REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG)
-        )
-        remote_image = registry.inspect(
-            REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG
-        )
+        self.teardown_cleanups.append((registry.rmi, REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG))
+        remote_image = registry.inspect(REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG)
 
-        self.assertEqual(
-            local_image[0]["Id"],
-            remote_image[0]["Id"]
-        )
+        self.assertEqual(local_image[0]["Id"], remote_image[0]["Id"])
 
     def test_pull_nonexistent_image(self):
         """Verify that a client cannot pull nonexistent image from Pulp.
@@ -246,14 +243,9 @@ class PullContentTestCase(unittest.TestCase):
         2. Assert that error is occurred and nothing has been pulled.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            "inexistentimagename"
-        )
+        local_url = urljoin(self.cfg.get_base_url(), "inexistentimagename")
         with self.assertRaises(exceptions.CalledProcessError):
             registry.pull(local_url)
 
@@ -276,7 +268,6 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         * `Pulp #4460 <https://pulp.plan.io/issues/4460>`_
         """
         cls.cfg = config.get_config()
-        gen_token_signing_keys(cls.cfg)
 
         client_api = gen_container_client()
         cls.repositories_api = RepositoriesContainerApi(client_api)
@@ -297,9 +288,7 @@ class PullOnDemandContentTestCase(unittest.TestCase):
 
             # Step 2
             cls.remote = cls.remotes_api.create(gen_container_remote(policy="on_demand"))
-            cls.teardown_cleanups.append(
-                (cls.remotes_api.delete, cls.remote.pulp_href)
-            )
+            cls.teardown_cleanups.append((cls.remotes_api.delete, cls.remote.pulp_href))
 
             # Step 3
             sync_data = RepositorySyncURL(remote=cls.remote.pulp_href)
@@ -332,7 +321,7 @@ class PullOnDemandContentTestCase(unittest.TestCase):
             distribution = cls.distributions_api.read(created_resources[0])
             cls.distribution_with_repo_version = cls.distributions_api.read(distribution.pulp_href)
             cls.teardown_cleanups.append(
-                (cls.distributions_api.delete, cls.distribution_with_repo_version.pulp_href)
+                (cls.distributions_api.delete, cls.distribution_with_repo_version.pulp_href,)
             )
 
             # remove callback if everything goes well
@@ -353,20 +342,15 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         """
         # Get local checksums for content synced from remote registy
         checksums = [
-            content["digest"] for content
-            in get_content(self.repo.to_dict())[CONTAINER_CONTENT_NAME]
+            content["digest"]
+            for content in get_content(self.repo.to_dict())[CONTAINER_CONTENT_NAME]
         ]
 
         # Assert that at least one layer is synced from remote:latest
         # and the checksum matched with remote
         self.assertTrue(
-            any(
-                [
-                    result["blobSum"] in checksums
-                    for result in get_docker_hub_remote_blobsums()
-                ]
-            ),
-            "Cannot find a matching layer on remote registry."
+            any([result["blobSum"] in checksums for result in get_docker_hub_remote_blobsums()]),
+            "Cannot find a matching layer on remote registry.",
         )
 
     def test_pull_image_from_repository(self):
@@ -379,14 +363,9 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         5. Ensure image is deleted after the test.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            self.distribution_with_repo.base_path
-        )
+        local_url = urljoin(self.cfg.get_base_url(), self.distribution_with_repo.base_path)
 
         registry.pull(local_url)
         self.teardown_cleanups.append((registry.rmi, local_url))
@@ -395,10 +374,7 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         registry.pull(REPO_UPSTREAM_NAME)
         remote_image = registry.inspect(REPO_UPSTREAM_NAME)
 
-        self.assertEqual(
-            local_image[0]["Id"],
-            remote_image[0]["Id"]
-        )
+        self.assertEqual(local_image[0]["Id"], remote_image[0]["Id"])
 
         new_artifact_count = self.artifacts_api.list().count
         self.assertGreater(new_artifact_count, self.artifact_count)
@@ -414,14 +390,9 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         4. Ensure image is deleted after the test.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            self.distribution_with_repo_version.base_path
-        )
+        local_url = urljoin(self.cfg.get_base_url(), self.distribution_with_repo_version.base_path)
 
         registry.pull(local_url)
         self.teardown_cleanups.append((registry.rmi, local_url))
@@ -430,10 +401,7 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         registry.pull(REPO_UPSTREAM_NAME)
         remote_image = registry.inspect(REPO_UPSTREAM_NAME)
 
-        self.assertEqual(
-            local_image[0]["Id"],
-            remote_image[0]["Id"]
-        )
+        self.assertEqual(local_image[0]["Id"], remote_image[0]["Id"])
         registry.rmi(REPO_UPSTREAM_NAME)
 
     def test_pull_image_with_tag(self):
@@ -445,28 +413,19 @@ class PullOnDemandContentTestCase(unittest.TestCase):
         4. Ensure image is deleted after the test.
         """
         registry = cli.RegistryClient(self.cfg)
-        registry.raise_if_unsupported(
-            unittest.SkipTest, "Test requires podman/docker"
-        )
+        registry.raise_if_unsupported(unittest.SkipTest, "Test requires podman/docker")
 
-        local_url = urljoin(
-            self.cfg.get_content_host_base_url(),
-            self.distribution_with_repo.base_path
-        ) + REPO_UPSTREAM_TAG
+        local_url = (
+            urljoin(self.cfg.get_base_url(), self.distribution_with_repo.base_path)
+            + REPO_UPSTREAM_TAG
+        )
 
         registry.pull(local_url)
         self.teardown_cleanups.append((registry.rmi, local_url))
         local_image = registry.inspect(local_url)
 
         registry.pull(REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG)
-        self.teardown_cleanups.append(
-            (registry.rmi, REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG)
-        )
-        remote_image = registry.inspect(
-            REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG
-        )
+        self.teardown_cleanups.append((registry.rmi, REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG))
+        remote_image = registry.inspect(REPO_UPSTREAM_NAME + REPO_UPSTREAM_TAG)
 
-        self.assertEqual(
-            local_image[0]["Id"],
-            remote_image[0]["Id"]
-        )
+        self.assertEqual(local_image[0]["Id"], remote_image[0]["Id"])